fix: created card component and reused it in step & add new card modal
on card added set mode and change plan
